move ValueForFeatureProfileDictionary from data_structures.py to __init__.py
