+++ langworld_db_data/tools/featureprofiles/feature_profile_reader.py
@@ -11,7 +11,7 @@ from langworld_db_data.constants.literals import (
 )
 from langworld_db_data.constants.paths import FEATURE_PROFILES_DIR
 from langworld_db_data.tools.common.files.csv_xls import read_dicts_from_csv
-from langworld_db_data.tools.featureprofiles.data_structures import (
+from langworld_db_data.tools.featureprofiles import (
     ValueForFeatureProfileDictionary,
 )
