+++ tests/test_tools_featureprofiles_feature_profile_reader.py
@@ -1,6 +1,6 @@
 import pytest
 
-from langworld_db_data.tools.featureprofiles.data_structures import (
+from langworld_db_data.tools.featureprofiles import (
     ValueForFeatureProfileDictionary,
 )
 from langworld_db_data.tools.featureprofiles.feature_profile_reader import FeatureProfileReader
